Switch to QNEthernet. 16 sample buffer.
The main issue seems to be jitter on packet arrival; receiving at intervals with extremes of  ~5 µs to ~2000 µs; and that's irrespective of whether using NativeEthernet or QNEthernet. Wireshark says packets are being sent at reasonably regular intervals. The question is, is this the ethernet port on machine I'm developing on? The switch? The Teensy ethernet adaptor? The ethernet libraries?