Switch to QNEthernet. 16 sample buffer.
The main issue seems to be jitter on packet arrival; receiving at intervals with extremes of ~5 µs to >2000 µs; and that's irrespective of whether using NativeEthernet or QNEthernet. Wireshark says packets are being transmitted at reasonably regular intervals. The question is, is this the culprit the ethernet port on the machine I'm developing on? The switch? Teensy's ethernet adaptor? Teensy's ethernet libraries?